% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{launchMorphoSimShiny}
\alias{launchMorphoSimShiny}
\title{Run the app in a web browser}
\usage{
launchMorphoSimShiny(inbrowser = TRUE)
}
\arguments{
\item{inbrowser}{Launch the app inside the system's default browser (default TRUE) or not.}
}
\description{
Run the app in a web browser
}
\examples{
if (interactive()) {
  launchMorphoSimShiny()
}

}
